requirements.txt
mysql-replication==0.21
loguru==0.4.1
pyyaml==5.3
envyaml==0.1912
pandas==0.25.3
sentry-sdk==0.19.1
prometheus-client==0.12.0
singlestoredb==0.4.0
